Unstable methods should not appear in runtime or d.ts (#4957)
