+++ cli/js/deno.ts
@@ -11,7 +11,6 @@ export {
 export { build } from "./build.ts";
 export { chmodSync, chmod } from "./ops/fs/chmod.ts";
 export { chownSync, chown } from "./ops/fs/chown.ts";
-export { transpileOnly, compile, bundle } from "./compiler/api.ts";
 export { customInspect, inspect } from "./web/console.ts";
 export { copyFileSync, copyFile } from "./ops/fs/copy_file.ts";
 export {
@@ -20,8 +19,7 @@ export {
   DiagnosticItem,
   DiagnosticMessageChain,
 } from "./diagnostics.ts";
-export { chdir, cwd } from "./ops/fs/dir.ts";
-export { applySourceMap, formatDiagnostics } from "./ops/errors.ts";
+export { chdir } from "./ops/fs/dir.ts";
 export { errors } from "./errors.ts";
 export {
   File,
@@ -51,7 +49,6 @@ export {
   Closer,
   Seeker,
 } from "./io.ts";
-export { linkSync, link } from "./ops/fs/link.ts";
 export {
   makeTempDirSync,
   makeTempDir,
@@ -61,34 +58,8 @@ export {
 } from "./ops/fs/make_temp.ts";
 export { metrics, Metrics } from "./ops/runtime.ts";
 export { mkdirSync, mkdir, MkdirOptions } from "./ops/fs/mkdir.ts";
-export {
-  connect,
-  listen,
-  listenDatagram,
-  DatagramConn,
-  Listener,
-  Conn,
-  ShutdownMode,
-  shutdown,
-} from "./net.ts";
-export {
-  dir,
-  env,
-  exit,
-  execPath,
-  hostname,
-  loadavg,
-  osRelease,
-} from "./ops/os.ts";
-export {
-  permissions,
-  PermissionName,
-  PermissionState,
-  PermissionStatus,
-  Permissions,
-} from "./permissions.ts";
-export { openPlugin } from "./ops/plugins.ts";
-export { kill } from "./ops/process.ts";
+export { connect, listen, Listener, Conn } from "./net.ts";
+export { dir, env, exit, execPath, hostname } from "./ops/os.ts";
 export { run, RunOptions, Process, ProcessStatus } from "./process.ts";
 export { DirEntry, readDirSync, readDir } from "./ops/fs/read_dir.ts";
 export { readFileSync, readFile } from "./read_file.ts";
@@ -98,14 +69,10 @@ export { realPathSync, realPath } from "./ops/fs/real_path.ts";
 export { removeSync, remove, RemoveOptions } from "./ops/fs/remove.ts";
 export { renameSync, rename } from "./ops/fs/rename.ts";
 export { resources, close } from "./ops/resources.ts";
-export { signal, signals, Signal, SignalStream } from "./signals.ts";
 export { FileInfo, statSync, lstatSync, stat, lstat } from "./ops/fs/stat.ts";
-export { symlinkSync, symlink } from "./ops/fs/symlink.ts";
-export { connectTls, listenTls, startTls } from "./tls.ts";
+export { connectTls, listenTls } from "./tls.ts";
 export { truncateSync, truncate } from "./ops/fs/truncate.ts";
-export { isatty, setRaw } from "./ops/tty.ts";
-export { umask } from "./ops/fs/umask.ts";
-export { utimeSync, utime } from "./ops/fs/utime.ts";
+export { isatty } from "./ops/tty.ts";
 export { version } from "./version.ts";
 export { writeFileSync, writeFile, WriteFileOptions } from "./write_file.ts";
 export { writeTextFileSync, writeTextFile } from "./write_text_file.ts";
