+++ cli/js/globals_unstable.ts
@@ -1,0 +1,5 @@
+// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+
+export const unstableMethods = {};
+
+export const unstableProperties = {};
