+++ cli/js/net.ts
@@ -48,6 +48,7 @@ export class ConnImpl implements Conn {
     close(this.rid);
   }
 
+  // TODO(lucacasonato): make this unavailable in stable
   closeWrite(): void {
     netOps.shutdown(this.rid, netOps.ShutdownMode.Write);
   }
@@ -141,57 +142,20 @@ export interface Conn extends Reader, Writer, Closer {
 export interface ListenOptions {
   port: number;
   hostname?: string;
-  transport?: "tcp" | "udp";
-}
-
-export interface UnixListenOptions {
-  transport: "unix" | "unixpacket";
-  path: string;
+  transport?: "tcp";
 }
 
 export function listen(
   options: ListenOptions & { transport?: "tcp" }
 ): Listener;
-export function listen(
-  options: UnixListenOptions & { transport: "unix" }
-): Listener;
-export function listen(options: ListenOptions | UnixListenOptions): Listener {
-  let res;
-
-  if (options.transport === "unix") {
-    res = netOps.listen(options);
-  } else {
-    res = netOps.listen({
-      transport: "tcp",
-      hostname: "127.0.0.1",
-      ...(options as ListenOptions),
-    });
-  }
+export function listen(options: ListenOptions): Listener {
+  const res = netOps.listen({
+    transport: "tcp",
+    hostname: "127.0.0.1",
+    ...(options as ListenOptions),
+  });
 
   return new ListenerImpl(res.rid, res.localAddr);
-}
-
-export function listenDatagram(
-  options: ListenOptions & { transport: "udp" }
-): DatagramConn;
-export function listenDatagram(
-  options: UnixListenOptions & { transport: "unixpacket" }
-): DatagramConn;
-export function listenDatagram(
-  options: ListenOptions | UnixListenOptions
-): DatagramConn {
-  let res;
-  if (options.transport === "unixpacket") {
-    res = netOps.listen(options);
-  } else {
-    res = netOps.listen({
-      transport: "udp",
-      hostname: "127.0.0.1",
-      ...(options as ListenOptions),
-    });
-  }
-
-  return new DatagramImpl(res.rid, res.localAddr);
 }
 
 export interface ConnectOptions {
