+++ cli/ops/runtime.rs
@@ -24,18 +24,19 @@ fn op_start(
 
   Ok(JsonOp::Sync(json!({
     // TODO(bartlomieju): `cwd` field is not used in JS, remove?
-    "cwd": &env::current_dir().unwrap(),
-    "pid": std::process::id(),
     "args": gs.flags.argv.clone(),
-    "repl": gs.flags.subcommand == DenoSubcommand::Repl,
-    "location": state.main_module.to_string(),
+    "cwd": &env::current_dir().unwrap(),
     "debugFlag": gs.flags.log_level.map_or(false, |l| l == log::Level::Debug),
-    "versionFlag": gs.flags.version,
-    "v8Version": version::v8(),
     "denoVersion": version::DENO,
-    "tsVersion": version::TYPESCRIPT,
+    "location": state.main_module.to_string(),
     "noColor": !colors::use_color(),
+    "pid": std::process::id(),
+    "repl": gs.flags.subcommand == DenoSubcommand::Repl,
     "target": env!("TARGET"),
+    "tsVersion": version::TYPESCRIPT,
+    "unstableFlag": gs.flags.unstable,
+    "v8Version": version::v8(),
+    "versionFlag": gs.flags.version,
   })))
 }
