--- cli/tests/unstable.out
@@ -1,1 +1,0 @@
-Unstable API 'Deno.openPlugin'. The --unstable flag must be provided.
