+++ cli/tests/unstable_disabled.out
@@ -1,0 +1,5 @@
+[WILDCARD]
+error TS2339: Property 'loadavg' does not exist on type 'typeof Deno'.
+console.log(Deno.loadavg);
+                 ~~~~~~~
+    at [WILDCARD]/cli/tests/unstable.ts:1:18
