+++ cli/tests/unstable_disabled_js.out
@@ -1,0 +1,1 @@
+undefined
