+++ cli/tests/unstable_enabled.out
@@ -1,0 +1,1 @@
+[Function: loadavg]
