+++ cli/tests/unstable_enabled_js.out
@@ -1,0 +1,1 @@
+[Function: loadavg]
